EP -- zadatak 01/requirements.txt
autopep8==2.0.2
paho-mqtt==1.6.1
pycodestyle==2.10.0
tomli==2.0.1
